fix: isSupporterPlus calls simplified and placed together
