+++ support-frontend/assets/components/paymentButton/defaultPaymentButtonContainer.tsx
@@ -1,10 +1,10 @@
 import type { ContributionType } from 'helpers/contributions';
 import { simpleFormatAmount } from 'helpers/forms/checkouts';
 import { currencies } from 'helpers/internationalisation/currency';
+import { isSupporterPlus } from 'helpers/redux/checkout/product/selectors/isSupporterPlus';
 import { getContributionType } from 'helpers/redux/checkout/product/selectors/productType';
 import { getUserSelectedAmount } from 'helpers/redux/checkout/product/selectors/selectedAmount';
 import { useContributionsSelector } from 'helpers/redux/storeHooks';
-import { shouldShowSupporterPlusMessaging } from 'helpers/supporterPlus/showMessaging';
 import { showThreeTierCheckout } from 'pages/supporter-plus-landing/setup/threeTierABTest';
 import { DefaultPaymentButton } from './defaultPaymentButton';
 
@@ -63,7 +63,7 @@ export function DefaultPaymentButtonContainer({
 
 	const testId = 'qa-contributions-landing-submit-contribution-button';
 
-	const amountIsAboveThreshold = shouldShowSupporterPlusMessaging(
+	const amountIsAboveThreshold = isSupporterPlus(
 		contributionType,
 		selectedAmounts,
 		otherAmounts,
